server_django/requirements.txt
django==3.2.18
djangorestframework-jwt==1.11.0
djangorestframework==3.14.0
mysqlclient==2.1.1
drf-yasg==1.21.5
django-imagekit
pilkit
python-decouple==3.3
django-phonenumber-field==7.1.0

protobuf==3.20.3

tensorflow==2.7.0
keras==2.13.1
scikit-learn==1.3.0
opencv-python==4.8.0.74
h5py==3.9.0
matplotlib==3.7.2
Pillow==9.5.0
requests==2.31.0
psutil==5.9.5
imageio==2.31.1
imutils==0.5.3
urllib3==2.0.2

django-crontab==0.7.1
celery==5.3.1
django-celery-results==2.5.0

